+++ CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Version 1.15.1 - 2025-07-xx
+## Version 1.15.1 - 2025-06-05
 
 ### Fixed
 
 - Paths with leading slash `/` are handled resiliently
+- Option `httpAgent` to pass http(s) agent to requests
 
 ## Version 1.15.0 - 2025-06-03